// apps/web-user/components/layout/header/types.ts
import type { ReactNode } from "react";

export type NavItem = {
  href: string;
  label: string;
  icon?: ReactNode;
  exact?: boolean;
  children?: Omit<NavItem, "children">[];
};

export interface HeaderProps {
  items: NavItem[];
  brand: string;
  brandHref: string;
  actions?: ReactNode;
  localeSwitcher?: ReactNode;
  authActions?: ReactNode;
}